services/telegram_bot/requirements.txt
flask
google-cloud-firestore
google-cloud-secret-manager
google-api-python-client
google-auth
google-cloud-storage
requests